Added Client-side Rendering for API calls, Fetched Full Details Page, Refactored Main Page Layout, Added State-management
Decided to change to client side api calls, as it vastly improved efficiency and enabled state management to work. Have all the details for each pokemon fetched from the API and formatted, just need to show on page and fix layout. Refactored the main page layout, so client components make the API calls, and added state to the key components here.